+++ src/frontend/apps/e2e/__tests__/app-impress/doc-header.spec.ts
@@ -69,6 +69,8 @@ test.describe('Doc Header', () => {
   }) => {
     await createDoc(page, 'doc-update', browserName, 1);
 
+    await writeInEditor({ page, text: 'Hello Content' });
+
     const card = page.getByLabel(
       'It is the card information about the document.',
     );
@@ -92,6 +94,9 @@ test.describe('Doc Header', () => {
       .click();
     await expect(
       page.getByRole('menuitem', { name: 'Download' }),
+    ).toBeVisible();
+    await expect(
+      page.getByRole('menuitem', { name: 'Word count: 2 words' }),
     ).toBeVisible();
   });